Add functionality allow the interface to be dynamically set.
